+++ vitest.config.ts
@@ -51,6 +51,10 @@ export default defineConfig({
             provider: 'v8',
             reporter: ['text', 'json-summary'],
         },
-        include: ['apps/**/*.test.ts', 'packages/**/*.test.ts'],
+        include: [
+            'apps/**/*.test.ts',
+            'packages/**/*.test.ts',
+            'tests/**/*.test.ts',
+        ],
     },
 });